split getServerSide props to its own, and add helper function to know where to send the user next
