+++ packages/applicant/src/utils/generateUrlForMandatoryQuestionNextPage.ts
@@ -1,0 +1,40 @@
+import { GrantMandatoryQuestionDto } from '../services/GrantMandatoryQuestionService';
+import { routes } from './routes';
+
+export const generateUrlForMandatoryQuestionNextPage = (
+  mandatoryQuestion: GrantMandatoryQuestionDto,
+  mandatoryQuestionId: string
+) => {
+  const redirect = { destination: '', permanent: false };
+  if (mandatoryQuestion.name) {
+    redirect.destination =
+      routes.mandatoryQuestions.addressPage(mandatoryQuestionId);
+  }
+  if (
+    !!mandatoryQuestion.addressLine1 &&
+    !!mandatoryQuestion.city &&
+    !!mandatoryQuestion.postcode
+  ) {
+    redirect.destination =
+      routes.mandatoryQuestions.typePage(mandatoryQuestionId);
+  }
+  if (mandatoryQuestion.orgType) {
+    redirect.destination =
+      routes.mandatoryQuestions.companiesHouseNumberPage(mandatoryQuestionId);
+  }
+  if (mandatoryQuestion.companiesHouseNumber) {
+    redirect.destination =
+      routes.mandatoryQuestions.charityCommissionNumberPage(
+        mandatoryQuestionId
+      );
+  }
+  if (mandatoryQuestion.charityCommissionNumber) {
+    redirect.destination =
+      routes.mandatoryQuestions.fundingAmountPage(mandatoryQuestionId);
+  }
+  if (mandatoryQuestion.fundingAmount) {
+    redirect.destination =
+      routes.mandatoryQuestions.fundingLocationPage(mandatoryQuestionId);
+  }
+  return redirect;
+};
